wip(api): continue with scaffolding and simple setup

+++ scripts/database/_wipe.ts
@@ -1,7 +1,22 @@
-import { db, schema } from "../../src/db";
+import { db, schema } from "@/db";
+import { logger } from "@/setup/logger";
 
-console.log("Wiping the database....");
-await db.delete(schema.user);
-await db.delete(schema.toolSnapshot);
-await db.delete(schema.tool);
-await db.delete(schema.category);
+if (
+    Bun.env.NODE_ENV !== "development" &&
+    Bun.env.NODE_ENV?.toString() !== "test"
+) {
+    throw new Error(
+        "This script must be run in development or test environment.",
+    );
+}
+
+logger.info("Wiping the database....");
+try {
+    await db.delete(schema.user);
+    await db.delete(schema.toolSnapshot);
+    await db.delete(schema.tool);
+    await db.delete(schema.category);
+    await db.delete(schema.categoryToolConnections);
+} catch (e) {
+    logger.error("Error wiping the database. Skipping...", e);
+}
